+++ client/src/App.js
@@ -6,7 +6,7 @@ import CustomHeader from "./components/customHeader"; //Include Heder
 import Home from "./pages/Home";
 import "./assets/css/content.css";
 import AdminJobs from "./pages/AdminJobs";
-import GradJobs from "./pages/GradJobs";
+import GradJobs from "./pages/GradJobs/GradJobs";
 import VerifyCertificateNum from "../components/users/VerifyCertificateNum";
 import { Signup } from "./components/Pages/signup/Signup";
