+++ client/src/pages/Button/Button.css
@@ -17,7 +17,7 @@ button:text {
 }
 
 button:hover {
-	background-color: #790503;
+	background-color: #be0b08;
 }
 
 /* Float cancel and signup buttons and add an equal width */
